+++ packages/orq-rc/src/orq_ai_sdk/snippets.py
@@ -82,7 +82,12 @@ class Snippets(BaseSDK):
             return utils.unmarshal_json(
                 http_res.text, Optional[models.FindOnePromptSnippetResponseBody]
             )
-        if utils.match_response(http_res, ["400", "404", "4XX", "5XX"], "*"):
+        if utils.match_response(http_res, ["400", "404", "4XX"], "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise models.APIError(
                 "API error occurred", http_res.status_code, http_res_text, http_res
@@ -170,7 +175,12 @@ class Snippets(BaseSDK):
             return utils.unmarshal_json(
                 http_res.text, Optional[models.FindOnePromptSnippetResponseBody]
             )
-        if utils.match_response(http_res, ["400", "404", "4XX", "5XX"], "*"):
+        if utils.match_response(http_res, ["400", "404", "4XX"], "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError(
                 "API error occurred", http_res.status_code, http_res_text, http_res